+++ packages/isaacscript-cli/src/customStage.ts
@@ -2,14 +2,15 @@ import chalk from "chalk";
 import path from "path";
 import * as tstl from "typescript-to-lua";
 import xml2js from "xml2js";
+import { getJSONRoomDoorSlotFlags } from "./common";
 import { CWD, MOD_SOURCE_PATH } from "./constants";
 import { PackageManager } from "./enums/PackageManager";
 import * as file from "./file";
 import {
-  CustomStage,
+  CustomStageLua,
   CustomStageRoomMetadata,
   CustomStageTSConfig,
-} from "./interfaces/copied/CustomStage";
+} from "./interfaces/copied/CustomStageLua";
 import { JSONRoomsFile } from "./interfaces/copied/JSONRoomsFile";
 import { getPackageManagerAddCommand } from "./packageManager";
 import { getCustomStagesFromTSConfig } from "./tsconfig";
@@ -92,12 +93,12 @@ function validateMetadataLuaFileExists(
 
 /**
  * This parses all of the end-user's XML files and gathers metadata about all of the rooms within.
- * (In other words, this creates the full set of `CustomStage` objects.)
+ * (In other words, this creates the full set of `CustomStageLua` objects.)
  */
 async function getCustomStagesWithMetadata(
   customStagesTSConfig: CustomStageTSConfig[],
   verbose: boolean,
-): Promise<CustomStage[]> {
+): Promise<CustomStageLua[]> {
   if (!file.exists(METADATA_LUA_PATH, verbose)) {
     error(
       `${chalk.red(
@@ -106,70 +107,123 @@ async function getCustomStagesWithMetadata(
     );
   }
 
-  const customStages: CustomStage[] = [];
+  const customStagesLua: CustomStageLua[] = [];
 
   for (const customStageTSConfig of customStagesTSConfig) {
-    const xmlPath = path.resolve(CWD, customStageTSConfig.xmlPath);
-    if (!file.exists(xmlPath, verbose)) {
+    const { name } = customStageTSConfig;
+    if (name === "") {
       error(
-        `${chalk.red(
-          "Failed to find the custom stage XML file at:",
-        )} ${chalk.green(xmlPath)}`,
+        chalk.red(
+          "One of the custom stages has a blank name, which is not allowed.",
+        ),
       );
     }
 
-    const xmlContents = file.read(xmlPath, verbose);
+    const { xmlPath } = customStageTSConfig;
+    if (xmlPath === "") {
+      error(
+        chalk.red(
+          `The "${name}" custom stage has a blank "xmlPath" property, which is not allowed.`,
+        ),
+      );
+    }
+
+    const resolvedXMLPath = path.resolve(CWD, xmlPath);
+    if (!file.exists(resolvedXMLPath, verbose)) {
+      error(
+        `${chalk.red(
+          "Failed to find the custom stage XML file at:",
+        )} ${chalk.green(resolvedXMLPath)}`,
+      );
+    }
+
+    const xmlContents = file.read(resolvedXMLPath, verbose);
     // eslint-disable-next-line no-await-in-loop
     const jsonRoomsFile = (await xml2js.parseStringPromise(
       xmlContents,
     )) as JSONRoomsFile;
 
+    const roomVariantSet = new Set<number>();
     const customStageRoomsMetadata: CustomStageRoomMetadata[] = [];
 
     for (const room of jsonRoomsFile.rooms.room) {
-      const baseVariant = parseIntSafe(room.$.variant);
-      if (Number.isNaN(baseVariant)) {
+      const typeString = room.$.type;
+      const type = parseIntSafe(typeString);
+      if (Number.isNaN(type)) {
         error(
-          `Failed to parse the variant of one of the custom stage rooms: ${room.$.variant}`,
+          `Failed to parse the type of one of the "${name}" custom stage rooms: ${typeString}`,
         );
       }
+
+      const variantString = room.$.variant;
+      const baseVariant = parseIntSafe(variantString);
+      if (Number.isNaN(baseVariant)) {
+        error(
+          `Failed to parse the variant of one of the "${name}" custom stage rooms: ${variantString}`,
+        );
+      }
+
+      if (roomVariantSet.has(baseVariant)) {
+        error(
+          chalk.red(
+            `There is more than one room with a variant of "${baseVariant}" in the "${name}" custom stage. Make sure that each room has a unique variant. (The room variant is also called the "ID" in Basement Renovator.)`,
+          ),
+        );
+      }
+      roomVariantSet.add(baseVariant);
+
       const roomVariantPrefix =
         customStageTSConfig.roomVariantPrefix * ROOM_VARIANT_MULTIPLIER;
       const variant = roomVariantPrefix + baseVariant;
 
-      const shape = parseIntSafe(room.$.shape);
-      if (Number.isNaN(baseVariant)) {
+      const subTypeString = room.$.subtype;
+      const subType = parseIntSafe(subTypeString);
+      if (Number.isNaN(subType)) {
         error(
-          `Failed to parse the shape of one of the custom stage rooms: ${room.$.shape}`,
+          `Failed to parse the sub-type of one of the "${name}" custom stage rooms: ${subTypeString}`,
         );
       }
 
-      const weight = parseFloat(room.$.weight);
+      const shapeString = room.$.shape;
+      const shape = parseIntSafe(shapeString);
       if (Number.isNaN(baseVariant)) {
         error(
-          `Failed to parse the weight of one of the custom stage rooms: ${room.$.weight}`,
+          `Failed to parse the shape of one of the "${name}" custom stage rooms: ${shapeString}`,
+        );
+      }
+
+      const doorSlotFlags = getJSONRoomDoorSlotFlags(room);
+
+      const weightString = room.$.weight;
+      const weight = parseFloat(weightString);
+      if (Number.isNaN(baseVariant)) {
+        error(
+          `Failed to parse the weight of one of the "${name}" custom stage rooms: ${weightString}`,
         );
       }
 
       const customStageRoomMetadata: CustomStageRoomMetadata = {
+        type,
         variant,
+        subType,
         shape,
+        doorSlotFlags,
         weight,
       };
       customStageRoomsMetadata.push(customStageRoomMetadata);
     }
 
-    const customStage: CustomStage = {
+    const customStageLua: CustomStageLua = {
       ...customStageTSConfig,
       roomsMetadata: customStageRoomsMetadata,
     };
-    customStages.push(customStage);
+    customStagesLua.push(customStageLua);
   }
 
-  return customStages;
+  return customStagesLua;
 }
 
-function convertCustomStagesToLua(customStages: CustomStage[]): string {
+function convertCustomStagesToLua(customStages: CustomStageLua[]): string {
   const customStagesString = JSON.stringify(customStages);
   const fakeTypeScriptFile = `return ${customStagesString}`;
   const result = tstl.transpileString(fakeTypeScriptFile, {