+++ packages/isaacscript-common/src/features/customStage/init.ts
@@ -1,17 +1,73 @@
+import {
+  DoorSlotFlag,
+  RoomShape,
+  RoomType,
+} from "isaac-typescript-definitions";
 import { isArray } from "../../functions/array";
-import { CustomStage } from "../../interfaces/CustomStage";
+import { CustomStage, RoomTypeMap } from "../../interfaces/CustomStage";
+import {
+  CustomStageLua,
+  CustomStageRoomMetadata,
+} from "../../interfaces/CustomStageLua";
 import * as metadataJSON from "./metadata.json"; // This will correspond to "metadata.lua" at run-time.
 import { customStagesMap } from "./v";
 
 export function customStageInit(): void {
-  const customStages = metadataJSON as unknown as CustomStage[];
-  if (!isArray(customStages)) {
+  const customStagesLua = metadataJSON as unknown as CustomStageLua[];
+  if (!isArray(customStagesLua)) {
     error(
       'The IsaacScript standard library attempted to read the custom stage metadata from the "metadata.lua" file, but it was not an array.',
     );
   }
 
-  for (const customStage of customStages) {
+  for (const customStageLua of customStagesLua) {
+    const roomTypeMap = getRoomTypeMap(customStageLua);
+    const customStage: CustomStage = {
+      ...customStageLua,
+      roomTypeMap,
+    };
     customStagesMap.set(customStage.name, customStage);
   }
+}
+
+function getRoomTypeMap(customStageLua: CustomStageLua): RoomTypeMap {
+  const roomTypeMap = new Map<
+    RoomType,
+    Map<RoomShape, Map<DoorSlotFlag, CustomStageRoomMetadata[]>>
+  >();
+
+  for (const roomMetadata of customStageLua.roomsMetadata) {
+    const roomType = roomMetadata.type as RoomType;
+
+    let roomShapeMap = roomTypeMap.get(roomType);
+    if (roomShapeMap === undefined) {
+      roomShapeMap = new Map<
+        RoomShape,
+        Map<DoorSlotFlag, CustomStageRoomMetadata[]>
+      >();
+      roomTypeMap.set(roomType, roomShapeMap);
+    }
+
+    const roomShape = roomMetadata.shape as RoomShape;
+
+    let roomDoorSlotFlagMap = roomShapeMap.get(roomShape);
+    if (roomDoorSlotFlagMap === undefined) {
+      roomDoorSlotFlagMap = new Map<
+        BitFlags<DoorSlotFlag>,
+        CustomStageRoomMetadata[]
+      >();
+      roomShapeMap.set(roomShape, roomDoorSlotFlagMap);
+    }
+
+    const doorSlotFlags = roomMetadata.doorSlotFlags as BitFlags<DoorSlotFlag>;
+    let rooms = roomDoorSlotFlagMap.get(doorSlotFlags);
+    if (rooms === undefined) {
+      rooms = [];
+      roomDoorSlotFlagMap.set(doorSlotFlags, rooms);
+    }
+
+    rooms.push(roomMetadata);
+  }
+
+  return roomTypeMap;
 }