+++ packages/isaacscript-common/src/features/customStage/v.ts
@@ -1,3 +1,7 @@
 import { CustomStage } from "../../interfaces/CustomStage";
 
+/** Indexed by custom stage name. */
 export const customStagesMap = new Map<string, CustomStage>();
+
+/** Indexed by room variant. */
+export const customStageCachedRoomData = new Map<int, Readonly<RoomConfig>>();